// app/modules/enums/CountyIdType.ts
export enum CountyIdType {
  TAIPEI = 'A',
  TAICHUNG = 'B',
  KEELUNG = 'C',
  TAINAN = 'D',
  KAOHSIUNG = 'E',
  NEW_TAIPEI = 'F',
  YILAN_COUNTY = 'G',
  TAOYUAN = 'H',
  CHIAYI = 'I',
  HSINCHU_COUNTY = 'J',
  MIAOLI_COUNTY = 'K',
  NANTOU_COUNTY = 'M',
  CHANGHUA_COUNTY = 'N',
  HSINCHU = 'O',
  YUNLIN_COUNTY = 'P',
  CHIAYI_COUNTY = 'Q',
  PINGTUNG_COUNTY = 'T',
  HUALIEN_COUNTY = 'U',
  TAITUNG_COUNTY = 'V',
  KINMEN_COUNTY = 'W',
  PENGHU_COUNTY = 'X',
  LIENCHIANG_COUNTY = 'Z'
}